backend/requirements.txt
fastapi
uvicorn
livekit-api
pydantic-settings
python-dotenv
